JsonBuilder - accept UTF-16 and UTF-32 names
Allow JsonBuilder to accept UTF-16 and UTF-32 name parameters to
push_back, push_front, and AddValue methods. The name is converted to
UTF-8 before being stored, so the value's `Name` property is always
returned as UTF-8.

- Old: push_back's name parameter is a `std::string_view`, which is
  treated as a UTF-8 string.
- New: push_back's name parameter is anything that can be converted into
  a `std::basic_string_view<CH>`, where `CH` can be any trivial type of
  size 1 (treated as UTF-8), 2 (treated as UTF-16), or 4 (treated as
  UTF-32).

Fancy template stuff was needed because there are 5 different
frequently-used string_builder types:

- std::string_view
- std::u8string_view (new in C++20, so supporting it explicitly would be
  a pain).
- std::u16string_view
- std::u32string_view
- std::wstring_view (char may be 16 or 32 bit, so supporting it
  explicitly would be a pain).

I didn't want to write 5 overloads for each of the 6 methods (30 methods
total), and I didn't want to explicitly support u8string_view and
wstring_view, so templates seemed better.

JsonBuilder.h:

- Add some template metaprogramming helpers that determine whether a
  given type can be converted into a `basic_string_view<CH>` with a
  valid `CH`, and whether `CH` should be treated as UTF-8, UTF-16, or
  UTF-32.
- For cosmetic reasons, move the declaration of the Validator class into
  JsonBuilder.cpp.
- For each of the 6 methods that needs to accept various kinds of
  strings, change the `name` parameter from `std::string_view` to
  template type `NameStringView`. Implement these as a conversion to
  `basic_string_view` followed by a call to new method `AddValueImpl`.
- Add `_newValueCommit` method to be called by the type helpers.
- Split the AddValue process into an `Init` step (different depending on
  string type) and a `Commit` step (different depending on data type).
- Type helpers are no longer involved in handling the name.
- Optimization: Change most of the type helpers to take the data by
  value instead of by reference.

Tricky bit of the implementation was how to split the AddValue process
into two steps while maintaining efficiency and behavior. In particular,
AddValue needs to support receiving name or data parameters that point
into the JsonBuilder, which is tricky if the JsonBuilder gets
reallocated.

TODO (next pr): Support AddValue with UTF-16 or UTF-32 data (convert to
UTF-8 as the value is added).