JsonBuilder - convert string values to UTF-8 (#27)
* JsonBuilder - convert string values to UTF-8

Add support for converting data from UTF-NN to UTF-8 when the data is
added to the JsonBuilder, i.e. you can now `push_back` a UTF-16 or
UTF-32 string and it will be converted to UTF-8 as part of the
`push_back` process.

Alternative design would be to add new data types for JsonUtf16 and
JsonUtf32, and then store the string in the JsonBuilder using its
original encoding. This seems problematic -- anything reading string
values would need to check the string's encoding and have separate code
paths for each of the 3 encodings.

* Also support converting latin1 and cp1252 input.